Proof-of-concept: Use Chrome DevTools Protocol Accessibility Tree
changelog: Internal, Automated Testing, Use Chrome DevTools Protocol Accessibility Tree for accessibility tests
